feat: :sparkles: extended script to generate schemas for 3rd party resources
